tests/wpt/web-platform-tests/feature-policy/experimental-features/lazyload/lazyload-disabled-tentative.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/feature-policy/experimental-features/resources/common.js"></script>
<style>
html, body {
  height: 100%;
  width: 100%;
}

iframe {
  width: 400px;
  height: 400px;
  margin: 10px;
}

.spacer {
  width: 100%;
  height: 10000px;
}
</style>
<div class="spacer"></div>
<script>
  let cross_origin_url =
      "http://{{hosts[alt][www]}}:{{ports[http][0]}}/" +
      "feature-policy/experimental-features/resources/lazyload-contents.html";

  let load_timeout = 600; // ms
  let expected_timeout_msg = false;

  window.scrollTo(0, 0);

  // Sanity-check: Make sure loading='lazy' works as intended.
  promise_test(async(t) => {
    // Add a frame with loading="lazy".
    let frame_lazy = createIframe(document.body, {
        id: "LAZY",
        loading: "lazy",
        src: `${cross_origin_url}?id=LAZY`
      });
    // Sanity-check: The frame is not visible.
    assert_greater_than(
        frame_lazy.getBoundingClientRect().top,
        window.innerHeight * 2,
        "Unexpected position for <iframe> with ID 'LAZY'.");
    let msg_or_timeout_attr_lazy =
        await waitForMessageOrTimeout(t, "LAZY", load_timeout);
    assert_equals(msg_or_timeout_attr_lazy,
                  expected_timeout_msg,
                  "With loading='lazy', the frame should not load.");
  }, "Sanity-check: Contents do not load immediately (no eager-loading) " +
     "when the loading attribute is 'lazy' and frame is in viewport.");


  // Verify that when 'lazyload' policy is disabled,  loading='eager' and
  // loading='auto' behave the same.
  promise_test(async(t) => {
    // Add a frame with loading="eager".
    let frame_lazy = createIframe(document.body, {
        id: "EAGER",
        loading: "eager",
        src: `${cross_origin_url}?id=EAGER`
      });
    // Sanity-check: The frame is not visible.
    assert_greater_than(
        frame_lazy.getBoundingClientRect().top,
        window.innerHeight * 2,
        "Unexpected position for <iframe> with ID 'EAGER'.");
    let msg_or_timeout_attr_eager =
        await waitForMessageOrTimeout(t, "EAGER", load_timeout);

    // Now do the same for loading='auto'.
    let frame_auto = createIframe(document.body, {
        id: "AUTO",
        loading: "auto",
        src: `${cross_origin_url}?id=AUTO`
    });
    // Sanity-check: The frame is not visible.
    assert_greater_than(
        frame_lazy.getBoundingClientRect().top,
        window.innerHeight * 2,
        "Unexpected position for <iframe> with ID 'AUTO'.");
    let msg_or_timeout_attr_auto =
        await waitForMessageOrTimeout(t, "AUTO", load_timeout);

    // The result should be the same.
    assert_equals(
          msg_or_timeout_attr_eager,
          msg_or_timeout_attr_auto,
          "loading='eager' not treated as 'auto'.");
  }, "When 'lazyload' feature is disabled, a frame cannot avoid lazyloading " +
     "by setting 'loading' attribute to 'eager'");
</script>
